feat: 更新了内网映射功能(OpenFrp支持)\nfeat: 现在使用OpenFrpc

+++ MSL/controls/Functions.cs
@@ -9,68 +9,49 @@ namespace MSL.controls
 {
     internal class Functions
     {
-        public static string Get(string path)
+        public static string Get(string path, string customUrl = "")
         {
-            string url;
-            if (MainWindow.serverLink != "https://msl.waheal.top")
+            string url = "https://api.waheal.top";
+            if (customUrl == "") { if (MainWindow.serverLink != "https://msl.waheal.top") url = MainWindow.serverLink + ":5000"; }
+            else url = customUrl;
+            WebClient webClient = new WebClient
             {
-                url = MainWindow.serverLink + ":5000";
-            }
-            else
-            {
-                url = "https://api.waheal.top";
-            }
-            WebClient webClient = new WebClient();
-            webClient.Credentials = CredentialCache.DefaultCredentials;
+                Credentials = CredentialCache.DefaultCredentials
+            };
             byte[] pageData = webClient.DownloadData(url + "/" + path);
             return Encoding.UTF8.GetString(pageData);
         }
 
-        public static string Post(string path,int contentType=0, string parameterData="",string customUrl="")
+        public static string Post(string path, int contentType = 0, string parameterData = "", string customUrl = "")
         {
-            string url;
-            if (customUrl == "")
-            {
-                if (MainWindow.serverLink != "https://msl.waheal.top")
-                {
-                    url = MainWindow.serverLink + ":5000";
-                }
-                else
-                {
-                    url = "https://api.waheal.top";
-                }
-            }
-            else
-            {
-                url = customUrl;
-            }
-            HttpWebRequest myRequest = (HttpWebRequest)WebRequest.Create(url+"/"+path);
+            string url = "https://api.waheal.top";
+            if (customUrl == "") { if (MainWindow.serverLink != "https://msl.waheal.top") url = MainWindow.serverLink + ":5000"; }
+            else url = customUrl;
+            HttpWebRequest myRequest = (HttpWebRequest)WebRequest.Create(url + "/" + path);
             byte[] buf = Encoding.GetEncoding("UTF-8").GetBytes(parameterData);
 
-            if (contentType == 0)
+            switch (contentType)
             {
-                myRequest.Method = "POST";
-                myRequest.Accept = "application/json";
-                myRequest.ContentType = "application/json; charset=UTF-8";
-                myRequest.ContentLength = buf.Length;
-                myRequest.MaximumAutomaticRedirections = 1;
-                myRequest.AllowAutoRedirect = true;
-            }
-            else if (contentType == 1)
-            {
-                myRequest.Method = "POST";
-                myRequest.Accept = "text/plain";
-                myRequest.ContentType = "text/plain; charset=UTF-8";
-                myRequest.ContentLength = buf.Length;
-                myRequest.MaximumAutomaticRedirections = 1;
-                myRequest.AllowAutoRedirect = true;
+                case 0:
+                    myRequest.Method = "POST";
+                    myRequest.Accept = "application/json";
+                    myRequest.ContentType = "application/json; charset=UTF-8";
+                    myRequest.ContentLength = buf.Length;
+                    myRequest.MaximumAutomaticRedirections = 1;
+                    myRequest.AllowAutoRedirect = true;
+                    break;
+                case 1:
+                    myRequest.Method = "POST";
+                    myRequest.Accept = "text/plain";
+                    myRequest.ContentType = "text/plain; charset=UTF-8";
+                    myRequest.ContentLength = buf.Length;
+                    myRequest.MaximumAutomaticRedirections = 1;
+                    myRequest.AllowAutoRedirect = true;
+                    break;
             }
 
             // 发送请求
-            using (Stream stream = myRequest.GetRequestStream())
-            {
-                stream.Write(buf, 0, buf.Length);
-            }
+            using (Stream stream = myRequest.GetRequestStream()) { stream.Write(buf, 0, buf.Length); }
 
             // 获取响应
             using (HttpWebResponse myResponse = (HttpWebResponse)myRequest.GetResponse())
@@ -84,35 +65,19 @@ namespace MSL.controls
 
         public static Tuple<int, int, int, int> VersionCompare(string version)
         {
-            if (version.StartsWith("*"))
-            {
-                return Tuple.Create(100, 100, 100, 100);
-            }
-
+            if (version.StartsWith("*")) return Tuple.Create(100, 100, 100, 100);
             // 使用正则表达式从版本号中提取主要版本号
             Regex regex = new Regex(@"(\d+(\.\d+)+)");
             Match match = regex.Match(version);
-            if (match.Success)
-            {
-                version = match.Groups[1].Value;
-            }
+            if (match.Success) version = match.Groups[1].Value;
 
             // 将版本号中的每个部分转换为整数，并进行比较
             string[] versionParts = version.Split('.');
             List<int> versionIntParts = new List<int>();
-            foreach (string part in versionParts)
-            {
-                if (int.TryParse(part, out int parsedPart))
-                {
-                    versionIntParts.Add(parsedPart);
-                }
-            }
+            foreach (string part in versionParts) { if (int.TryParse(part, out int parsedPart)) versionIntParts.Add(parsedPart); }
 
             // 添加0，以便对不完整的版本号进行比较（如1.7）
-            while (versionIntParts.Count < 4)
-            {
-                versionIntParts.Add(0);
-            }
+            while (versionIntParts.Count < 4) { versionIntParts.Add(0); }
 
             return Tuple.Create(versionIntParts[0], versionIntParts[1], versionIntParts[2], versionIntParts[3]);
         }
@@ -124,14 +89,8 @@ namespace MSL.controls
                 if (!Directory.Exists(destPath))
                 {
                     //目标目录不存在则创建
-                    try
-                    {
-                        Directory.CreateDirectory(destPath);
-                    }
-                    catch (Exception ex)
-                    {
-                        throw new Exception("创建目标目录失败：" + ex.Message);
-                    }
+                    try { _ = Directory.CreateDirectory(destPath); }
+                    catch (Exception ex) { throw new Exception("创建目标目录失败：" + ex.Message); }
                 }
                 //获得源文件下所有文件
                 List<string> files = new List<string>(Directory.GetFiles(sourcePath));
@@ -139,10 +98,7 @@ namespace MSL.controls
                 {
                     string destFile = Path.Combine(new string[] { destPath, Path.GetFileName(c) });
                     //覆盖模式
-                    if (File.Exists(destFile))
-                    {
-                        File.Delete(destFile);
-                    }
+                    if (File.Exists(destFile)) File.Delete(destFile);
                     File.Move(c, destFile);
                 });
                 //获得源文件下所有目录文件
@@ -159,10 +115,7 @@ namespace MSL.controls
                 });
                 Directory.Delete(sourcePath);
             }
-            else
-            {
-                throw new DirectoryNotFoundException("源目录不存在！");
-            }
+            else throw new DirectoryNotFoundException("源目录不存在");
         }
     }
 }
